refactor: Base 추가
