+++ src/test/java/nextstep/subway/acceptance/common/ProviderBase.java
@@ -1,0 +1,15 @@
+package nextstep.subway.acceptance.common;
+
+import static org.assertj.core.api.Assertions.*;
+
+import org.springframework.http.HttpStatus;
+
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+
+public abstract class ProviderBase {
+
+	protected static void 응답코드_검증(ExtractableResponse<Response> response, HttpStatus httpStatus) {
+		assertThat(response.statusCode()).isEqualTo(httpStatus.value());
+	}
+}
