+++ src/test/java/nextstep/subway/acceptance/station/StationProvider.java
@@ -1,15 +1,13 @@
 package nextstep.subway.acceptance.station;
 
 import static nextstep.subway.acceptance.station.StationRestAssuredProvider.*;
-import static org.assertj.core.api.Assertions.*;
 import static org.springframework.http.HttpStatus.*;
-
-import org.springframework.http.HttpStatus;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.subway.acceptance.common.ProviderBase;
 
-public class StationProvider {
+public class StationProvider extends ProviderBase {
 
 	public static ExtractableResponse<Response> 지하철역_생성_성공(String 역이름) {
 		ExtractableResponse<Response> response = 지하철역_생성(역이름);
@@ -26,10 +24,6 @@ public class StationProvider {
 	public static void 지하철역_제거_성공(String id) {
 		ExtractableResponse<Response> response = 지하철역_제거(id);
 		응답코드_검증(response, NO_CONTENT);
-	}
-
-	private static void 응답코드_검증(ExtractableResponse<Response> response, HttpStatus httpStatus) {
-		assertThat(response.statusCode()).isEqualTo(httpStatus.value());
 	}
 
 	public static String 지하철역_Id_추출(ExtractableResponse<Response> response) {
